The underlying date and time structures have been changed from the BCL class `DateTime` to the NodaTime class `ZonedDateTime`.
Taking into account that if the NTP server poll fails the component will return the system date and time, a new property `TheAnswerIsReliable` has been added to query about the reliability of the answer, considering that the NTP server response is reliable, and the system response is not.

The unit tests have been modified accordingly.